+++ src/renderers/opengl/tr_image.c
@@ -84,7 +84,7 @@ void GL_TextureMode( const char *string ) {
 	int		i;
 	
 	mode = NULL;
-	for ( i = 0 ; i < ARRAY_LEN( modes ) ; i++ ) {
+	for ( i = 0 ; (size_t)i < ARRAY_LEN( modes ) ; i++ ) {
 		if ( !Q_stricmp( modes[i].name, string ) ) {
 			mode = &modes[i];
 			break;
@@ -256,7 +256,7 @@ static void ResampleTexture( unsigned *in, int inwidth, int inheight, unsigned *
 	unsigned	p2[MAX_TEXTURE_SIZE];
 	byte		*pix1, *pix2, *pix3, *pix4;
 
-	if ( outwidth > ARRAY_LEN( p1 ) )
+	if ( (size_t)outwidth > ARRAY_LEN( p1 ) )
 		ri.Error( ERR_DROP, "ResampleTexture: max width" );
 								
 	fracstep = inwidth * 0x10000 / outwidth;
@@ -1428,7 +1428,7 @@ void R_SetColorMappings( void ) {
 
 	shift = tr.overbrightBits;
 
-	for ( i = 0; i < ARRAY_LEN( s_gammatable ); i++ ) {
+	for ( i = 0; (size_t)i < ARRAY_LEN( s_gammatable ); i++ ) {
 		if ( g == 1.0f ) {
 			inf = i;
 		} else {
@@ -1444,7 +1444,7 @@ void R_SetColorMappings( void ) {
 		s_gammatable[i] = inf;
 	}
 
-	for ( i = 0; i < ARRAY_LEN( s_intensitytable ); i++ ) {
+	for ( i = 0; (size_t)i < ARRAY_LEN( s_intensitytable ); i++ ) {
 		j = i * r_intensity->value;
 		if ( j > 255 ) {
 			j = 255;
